moving common into data
