+++ test/core/phone_test.py
@@ -1,7 +1,7 @@
 import unittest
 from pathlib import Path
 
-from howl.core.phone import PhonePhrase, PronunciationDictionary
+from howl.data.common.phone import PhonePhrase, PronunciationDictionary
 
 
 class TestPhonePhrase(unittest.TestCase):
