+++ test/data/dataset/serialize_test.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 import torch
 
-from howl.core.metadata import AudioClipMetadata
+from howl.data.common.metadata import AudioClipMetadata
 from howl.data.dataset.dataset import AudioClipDataset, DatasetType
 from howl.data.dataset.example import AudioClipExample
 from howl.data.dataset.serialize import AudioDatasetWriter
