+++ test/data/tokenizer_test.py
@@ -1,6 +1,6 @@
 import unittest
 
-from howl.core.vocab import Vocab
+from howl.data.common.vocab import Vocab
 from howl.data.tokenizer import WakeWordTokenizer
