+++ training/run/stitch_vocab_samples.py
@@ -1,7 +1,7 @@
 import argparse
 from pathlib import Path
 
-from howl.core.vocab import Vocab
+from howl.data.common.vocab import Vocab
 from howl.data.dataset import AudioDatasetWriter, WakeWordDatasetLoader
 from howl.data.labeler import WordFrameLabeler
 from howl.data.stitcher import WordStitcher
